SRID 28992 rename name from FeatureTypeMapping to match database added filtered (need 6 more)
