+++ ipo-deegree-workspaces/VoortgangsRapportageNatuur/datasources/feature/gebiedbeheer_landelijk.xml
@@ -1,9 +1,9 @@
 <SQLFeatureStore xmlns="http://www.deegree.org/datasource/feature/sql" xmlns:imna="http://www.cds.nl/imna/beheer" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" configVersion="3.2.0" xsi:schemaLocation="http://www.deegree.org/datasource/feature/sql http://schemas.deegree.org/datasource/feature/sql/3.2.0/sql.xsd">
   <JDBCConnId>vrn</JDBCConnId>
-  <StorageCRS srid="-1" dim="2D">EPSG:28992</StorageCRS>
+  <StorageCRS srid="28992" dim="2D">EPSG:28992</StorageCRS>
   <GMLSchema>../../appschemas/landelijk/imna_vrn_beheer.xsd</GMLSchema>
 
-  <FeatureTypeMapping xmlns:imna="http://www.cds.nl/imna/beheer" name="imna:GebiedBeheerLandelijk" table="vrn.gebiedbeheer_landelijk">
+  <FeatureTypeMapping xmlns:imna="http://www.cds.nl/imna/beheer" name="imna:LandelijkGebiedBeheer" table="vrn.gebiedbeheer_landelijk">
       <FIDMapping>
         <Column name="id"/>
     </FIDMapping>
