+++ ipo-deegree-workspaces/VoortgangsRapportageNatuur/datasources/feature/gebiedbeheer_landelijk_filtered.xml
@@ -1,0 +1,5 @@
+<VRNFilterSQLFeatureStore xmlns="urn:cds-vrn:deegree-extension:filter">
+	<DelegateFeatureStoreId>gebiedbeheer_landelijk</DelegateFeatureStoreId>
+	<FilterBronHouderGeometry>true</FilterBronHouderGeometry>
+	<FilterGebruikerThemaAutorisatie>true</FilterGebruikerThemaAutorisatie>
+</VRNFilterSQLFeatureStore>
