+++ ipo-deegree-workspaces/VoortgangsRapportageNatuur/datasources/feature/gebiedinrichting_provinciaal.xml
@@ -1,10 +1,9 @@
 <SQLFeatureStore xmlns="http://www.deegree.org/datasource/feature/sql" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" configVersion="3.2.0"
 xsi:schemaLocation="http://www.deegree.org/datasource/feature/sql http://schemas.deegree.org/datasource/feature/sql/3.2.0/sql.xsd">
   <JDBCConnId>vrn</JDBCConnId>
-   <StorageCRS srid="-1" dim="2D">EPSG:28992</StorageCRS>
-   <!--<StorageCRS srid="28992" dim="2D">urn:ogc:def:crs:EPSG::28992</StorageCRS>-->
+   <StorageCRS srid="28992" dim="2D">EPSG:28992</StorageCRS>
   <GMLSchema>../../appschemas/provinciaal/imna_vrn_inrichting.xsd</GMLSchema>
-<FeatureTypeMapping xmlns:imna="http://www.cds.nl/imna/inrichting" table="vrn.gebiedinrichting_provinciaal" name="imna:GebiedInrichtingProvinciaal">
+<FeatureTypeMapping xmlns:imna="http://www.cds.nl/imna/inrichting" table="vrn.gebiedinrichting_provinciaal" name="imna:ProvinciaalGebiedInrichting">
       <FIDMapping>
         <Column name="id"/>
     </FIDMapping>
