+++ ipo-deegree-workspaces/VoortgangsRapportageNatuur/datasources/feature/gebiedverwerving_vastgesteld.xml
@@ -2,7 +2,7 @@
   <JDBCConnId>vrn</JDBCConnId>
   <StorageCRS srid="28992">EPSG:28992</StorageCRS>
   <GMLSchema>../../appschemas/tagged/imna_vrn_verwerving.xsd</GMLSchema>
-<FeatureTypeMapping xmlns:imna="http://www.cds.nl/imna/verwerving" table="vrn.gebiedverwerving_landelijk_tagged" name="imna:GebiedVerwervingVastgesteld">
+<FeatureTypeMapping xmlns:imna="http://www.cds.nl/imna/verwerving" table="vrn.gebiedverwerving_landelijk_tagged" name="imna:VastgesteldGebiedVerwerving">
       <FIDMapping>
         <Column name="id"/>
     </FIDMapping>
